Move FS-bases utility functions to internal/fs package.
